image_bot/pyproject.toml
[project]
name = "image_bot"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "aiohttp (>=3.11.13,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "tortoise-orm[psycopg] (>=0.24.2,<0.25.0)",
    "aerich[toml] (>=0.8.2,<0.9.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
]

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
aerich = "^0.8.2"

[tool.aerich]
tortoise_orm = "image_bot.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
